requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
alsa-lib=1.2.13=hb9d3cd8_0
aom=3.9.1=hac33072_0
appdirs=1.4.4=pypi_0
beartype=0.17.0=pypi_0
blosc=1.21.6=he440d0b_1
brotli=1.1.0=hb9d3cd8_2
brotli-bin=1.1.0=hb9d3cd8_2
brunsli=0.1=h9c3ff4c_0
bypy=1.8.4=pypi_0
bzip2=1.0.8=h4bc722e_7
c-blosc2=2.15.2=h3122c55_1
ca-certificates=2024.12.14=hbcca054_0
cairo=1.18.2=h3394656_1
certifi=2024.2.2=pypi_0
charls=2.4.2=h59595ed_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
colt5-attention=0.10.19=pypi_0
contourpy=1.3.1=py310h3788b33_0
cycler=0.12.1=pyhd8ed1ab_1
cyrus-sasl=2.1.27=h54b06d7_7
dav1d=1.2.1=hd590300_0
dbus=1.13.6=h5008d03_3
dill=0.3.8=pypi_0
docker-pycreds=0.4.0=pypi_0
double-conversion=3.3.0=h59595ed_0
einops=0.7.0=pypi_0
expat=2.6.4=h5888daf_0
filelock=3.13.1=pypi_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=h77eed37_3
fontconfig=2.15.0=h7e30c49_1
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.55.3=py310h89163eb_1
freetype=2.12.1=h267a509_2
fsspec=2024.2.0=pypi_0
giflib=5.2.2=hd590300_0
gitdb=4.0.11=pypi_0
gitpython=3.1.41=pypi_0
graphite2=1.3.13=h59595ed_1003
harfbuzz=9.0.0=hda332d3_1
huggingface-hub=0.20.3=pypi_0
icu=75.1=he02047a_0
idna=3.6=pypi_0
imagecodecs=2024.9.22=py310h78a9a29_2
imageio=2.36.1=pyh12aca89_1
importlib-metadata=8.5.0=pyha770c72_1
jinja2=3.1.3=pypi_0
jxrlib=1.1=hd590300_3
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.7=py310h3788b33_0
krb5=1.21.3=h659f571_0
lazy-loader=0.4=pyhd8ed1ab_2
lazy_loader=0.4=pyhd8ed1ab_2
lcms2=2.16=hb7c19ff_0
ld_impl_linux-64=2.43=h712a8e2_2
lerc=4.0.0=h27087fc_0
libaec=1.1.3=h59595ed_0
libavif16=1.1.1=h1909e37_2
libblas=3.9.0=26_linux64_openblas
libbrotlicommon=1.1.0=hb9d3cd8_2
libbrotlidec=1.1.0=hb9d3cd8_2
libbrotlienc=1.1.0=hb9d3cd8_2
libcblas=3.9.0=26_linux64_openblas
libclang-cpp19.1=19.1.6=default_hb5137d0_0
libclang13=19.1.6=default_h9c6a7e4_0
libcups=2.3.3=h4637d8d_4
libdeflate=1.23=h4ddbbb0_0
libdrm=2.4.124=hb9d3cd8_0
libedit=3.1.20240808=pl5321h7949ede_0
libegl=1.7.0=ha4b6fd6_2
libexpat=2.6.4=h5888daf_0
libffi=3.4.2=h7f98852_5
libgcc=14.2.0=h77fa898_1
libgcc-ng=14.2.0=h69a702a_1
libgfortran=14.2.0=h69a702a_1
libgfortran5=14.2.0=hd5240d6_1
libgl=1.7.0=ha4b6fd6_2
libglib=2.82.2=h2ff4ddf_0
libglvnd=1.7.0=ha4b6fd6_2
libglx=1.7.0=ha4b6fd6_2
libgomp=14.2.0=h77fa898_1
libhwy=1.1.0=h00ab1b0_0
libiconv=1.17=hd590300_2
libjpeg-turbo=3.0.0=hd590300_1
libjxl=0.11.1=hdb8da77_0
liblapack=3.9.0=26_linux64_openblas
libllvm19=19.1.6=ha7bfdaf_0
liblzma=5.6.3=hb9d3cd8_1
libnsl=2.0.1=hd590300_0
libntlm=1.8=hb9d3cd8_0
libopenblas=0.3.28=pthreads_h94d23a6_1
libopengl=1.7.0=ha4b6fd6_2
libpciaccess=0.18=hd590300_0
libpng=1.6.45=h943b412_0
libpq=17.2=h3b95a9b_1
libsqlite=3.47.0=hadc24fc_1
libstdcxx=14.2.0=hc0a3c3a_1
libstdcxx-ng=14.2.0=h4852527_1
libtiff=4.7.0=hd9ff511_3
libuuid=2.38.1=h0b41bf4_0
libwebp-base=1.5.0=h851e524_0
libxcb=1.17.0=h8a09558_0
libxcrypt=4.4.36=hd590300_1
libxkbcommon=1.7.0=h2c5496b_1
libxml2=2.13.5=h8d12d68_1
libxslt=1.1.39=h76b75d6_0
libzlib=1.3.1=hb9d3cd8_2
libzopfli=1.0.3=h9c3ff4c_0
local-attention=1.9.0=pypi_0
lz4-c=1.10.0=h5888daf_1
markupsafe=2.1.5=pypi_0
matplotlib=3.10.0=py310hff52083_0
matplotlib-base=3.10.0=py310h68603db_0
mpmath=1.3.0=pypi_0
multiprocess=0.70.16=pypi_0
munkres=1.1.4=pyh9f0ad1d_0
mysql-common=9.0.1=h266115a_4
mysql-libs=9.0.1=he0572af_4
ncurses=6.5=he02047a_1
networkx=3.2.1=pypi_0
numpy=1.26.4=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.19.3=pypi_0
nvidia-nvjitlink-cu12=12.3.101=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
openjpeg=2.5.3=h5fbd93e_0
openldap=2.6.9=he970967_0
openssl=3.4.0=h7b32b05_1
packaging=23.2=pypi_0
pcre2=10.44=hba22ea6_2
pillow=10.2.0=pypi_0
pip=24.3.1=pyh8b19718_0
piq=0.8.0=pypi_0
pixman=0.44.2=h29eaf8c_0
protobuf=4.25.2=pypi_0
psutil=5.9.8=pypi_0
ptflops=0.7.2.2=pypi_0
pthread-stubs=0.4=hb9d3cd8_1002
pyparsing=3.2.1=pyhd8ed1ab_0
pyside6=6.8.1=py310hfd10a26_0
python=3.10.15=h4a871b0_2_cpython
python-dateutil=2.9.0.post0=pyhff2d567_1
python_abi=3.10=5_cp310
pywavelets=1.8.0=py310hf462985_0
pyyaml=6.0.1=pypi_0
qhull=2020.2=h434a139_5
qt6-main=6.8.1=h9d28a51_0
rav1e=0.6.6=he8a937b_2
readline=8.2=h8228510_1
requests=2.31.0=pypi_0
requests-toolbelt=1.0.0=pypi_0
safetensors=0.4.2=pypi_0
scikit-image=0.25.0=py310h5eaa309_1
scipy=1.15.0=py310hfa6ec8c_1
sentry-sdk=1.40.5=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=75.3.0=pyhd8ed1ab_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
snappy=1.2.1=h8bd8927_1
svt-av1=2.3.0=h5888daf_0
sympy=1.12=pypi_0
tifffile=2024.12.12=pyhd8ed1ab_0
timm=0.9.12=pypi_0
tk=8.6.13=noxft_h4845f30_101
torch=2.1.0+cu118=pypi_0
torchaudio=2.1.0+cu118=pypi_0
torchvision=0.16.0+cu118=pypi_0
tornado=6.4.2=py310ha75aee5_0
tqdm=4.66.1=pypi_0
triton=2.1.0=pypi_0
typing-extensions=4.9.0=pypi_0
tzdata=2024b=hc8b5060_0
unicodedata2=15.1.0=py310ha75aee5_1
urllib3=2.2.0=pypi_0
wandb=0.16.3=pypi_0
wayland=1.23.1=h3e06ad9_0
wheel=0.45.0=pyhd8ed1ab_0
xcb-util=0.4.1=hb711507_2
xcb-util-cursor=0.1.5=hb9d3cd8_0
xcb-util-image=0.4.0=hb711507_2
xcb-util-keysyms=0.4.1=hb711507_0
xcb-util-renderutil=0.3.10=hb711507_0
xcb-util-wm=0.4.2=hb711507_0
xkeyboard-config=2.43=hb9d3cd8_0
xorg-libice=1.1.2=hb9d3cd8_0
xorg-libsm=1.2.5=he73a12e_0
xorg-libx11=1.8.10=h4f16b4b_1
xorg-libxau=1.0.12=hb9d3cd8_0
xorg-libxcomposite=0.4.6=hb9d3cd8_2
xorg-libxcursor=1.2.3=hb9d3cd8_0
xorg-libxdamage=1.1.6=hb9d3cd8_0
xorg-libxdmcp=1.1.5=hb9d3cd8_0
xorg-libxext=1.3.6=hb9d3cd8_0
xorg-libxfixes=6.0.1=hb9d3cd8_0
xorg-libxi=1.8.2=hb9d3cd8_0
xorg-libxrandr=1.5.4=hb9d3cd8_0
xorg-libxrender=0.9.12=hb9d3cd8_0
xorg-libxtst=1.2.5=hb9d3cd8_3
xorg-libxxf86vm=1.1.6=hb9d3cd8_0
xz=5.2.6=h166bdaf_0
zfp=1.0.1=h5888daf_2
zipp=3.21.0=pyhd8ed1ab_1
zlib-ng=2.2.3=h7955e40_0
zstd=1.5.6=ha6fb4c9_0
